+++ packages/warehouses/src/warehouseClients/SnowflakeWarehouseClient.ts
@@ -8,6 +8,7 @@ import {
     SupportedDbtAdapter,
     WarehouseConnectionError,
     WarehouseQueryError,
+    WarehouseResults,
 } from '@lightdash/common';
 import * as crypto from 'crypto';
 import {
@@ -171,11 +172,15 @@ export class SnowflakeWarehouseClient extends WarehouseBaseClient<CreateSnowflak
         } as ConnectionOptions; // force type because accessUrl property is not recognised
     }
 
-    async runQuery(
-        sqlText: string,
-        tags?: Record<string, string>,
-        timezone?: string,
-    ) {
+    async streamQuery(
+        sql: string,
+        streamCallback: (data: WarehouseResults) => void,
+        options: {
+            values?: any[];
+            tags?: Record<string, string>;
+            timezone?: string;
+        },
+    ): Promise<void> {
         let connection: Connection;
         try {
             connection = createConnection(this.connectionOptions);
@@ -201,13 +206,15 @@ export class SnowflakeWarehouseClient extends WarehouseBaseClient<CreateSnowflak
                     `ALTER SESSION SET WEEK_START = ${snowflakeStartOfWeekIndex};`,
                 );
             }
-            if (tags) {
+            if (options?.tags) {
                 await this.executeStatement(
                     connection,
-                    `ALTER SESSION SET QUERY_TAG = '${JSON.stringify(tags)}';`,
+                    `ALTER SESSION SET QUERY_TAG = '${JSON.stringify(
+                        options?.tags,
+                    )}';`,
                 );
             }
-            const timezoneQuery = timezone || 'UTC';
+            const timezoneQuery = options?.timezone || 'UTC';
             console.debug(
                 `Setting Snowflake session timezone to ${timezoneQuery}`,
             );
@@ -228,11 +235,12 @@ export class SnowflakeWarehouseClient extends WarehouseBaseClient<CreateSnowflak
                 `ALTER SESSION SET QUOTED_IDENTIFIERS_IGNORE_CASE = FALSE;`,
             );
 
-            const result = await this.executeStreamStatement(
+            await this.executeStreamStatement(
                 connection,
-                sqlText,
+                sql,
+                streamCallback,
+                options,
             );
-            return result;
         } catch (e) {
             throw new WarehouseQueryError(e.message);
         } finally {
@@ -250,19 +258,35 @@ export class SnowflakeWarehouseClient extends WarehouseBaseClient<CreateSnowflak
     private async executeStreamStatement(
         connection: Connection,
         sqlText: string,
-    ) {
-        return new Promise<{
-            fields: Record<string, { type: DimensionType }>;
-            rows: any[];
-        }>((resolve, reject) => {
+        streamCallback: (data: WarehouseResults) => void,
+        options?: {
+            values?: any[];
+        },
+    ): Promise<void> {
+        return new Promise<void>((resolve, reject) => {
             connection.execute({
                 sqlText,
+                binds: options?.values,
                 streamResult: true,
                 complete: (err, stmt) => {
                     if (err) {
                         reject(new WarehouseQueryError(err.message));
                     }
-                    const rows: any[] = [];
+
+                    const columns = stmt.getColumns();
+                    const fields = columns
+                        ? columns.reduce(
+                              (acc, column) => ({
+                                  ...acc,
+                                  [column.getName()]: {
+                                      type: mapFieldType(
+                                          column.getType().toUpperCase(),
+                                      ),
+                                  },
+                              }),
+                              {},
+                          )
+                        : {};
 
                     pipeline(
                         stmt.streamRows(),
@@ -275,7 +299,7 @@ export class SnowflakeWarehouseClient extends WarehouseBaseClient<CreateSnowflak
                         new Writable({
                             objectMode: true,
                             write(chunk, encoding, callback) {
-                                rows.push(chunk);
+                                streamCallback({ fields, rows: [chunk] });
                                 callback();
                             },
                         }),
@@ -283,24 +307,7 @@ export class SnowflakeWarehouseClient extends WarehouseBaseClient<CreateSnowflak
                             if (error) {
                                 reject(new WarehouseQueryError(error.message));
                             } else {
-                                const columns = stmt.getColumns();
-                                const fields = columns
-                                    ? columns.reduce(
-                                          (acc, column) => ({
-                                              ...acc,
-                                              [column.getName()]: {
-                                                  type: mapFieldType(
-                                                      column
-                                                          .getType()
-                                                          .toUpperCase(),
-                                                  ),
-                                              },
-                                          }),
-                                          {},
-                                      )
-                                    : {};
-
-                                resolve({ fields, rows });
+                                resolve();
                             }
                         },
                     );
@@ -447,5 +454,55 @@ export class SnowflakeWarehouseClient extends WarehouseBaseClient<CreateSnowflak
             default:
                 return super.getMetricSql(sql, metric);
         }
+    }
+
+    async getTables(
+        schema?: string,
+        tags?: Record<string, string>,
+    ): Promise<WarehouseCatalog> {
+        const schemaFilter = schema ? `AND TABLE_SCHEMA ILIKE ?` : '';
+        const query = `
+            SELECT 
+                LOWER(TABLE_CATALOG) as "table_catalog", 
+                LOWER(TABLE_SCHEMA) as "table_schema",
+                LOWER(TABLE_NAME) as "table_name"
+            FROM information_schema.tables
+            WHERE TABLE_TYPE = 'BASE TABLE' 
+            ${schemaFilter}
+            ORDER BY 1,2,3
+        `;
+        const { rows } = await this.runQuery(
+            query,
+            tags,
+            undefined,
+            schema ? [schema] : undefined,
+        );
+        return this.parseWarehouseCatalog(rows, mapFieldType);
+    }
+
+    async getFields(
+        tableName: string,
+        schema?: string,
+        tags?: Record<string, string>,
+    ): Promise<WarehouseCatalog> {
+        const schemaFilter = schema ? `AND TABLE_SCHEMA ILIKE ?` : '';
+
+        const query = `
+            SELECT LOWER(TABLE_CATALOG) as "table_catalog",
+                   LOWER(TABLE_SCHEMA)  as "table_schema",
+                   LOWER(TABLE_NAME)    as "table_name",
+                   LOWER(COLUMN_NAME)   as "column_name",
+                   DATA_TYPE            as "data_type"
+            FROM information_schema.columns
+            WHERE TABLE_NAME ILIKE ? ${schemaFilter}
+            ORDER BY 1, 2, 3;
+        `;
+        const { rows } = await this.runQuery(
+            query,
+            tags,
+            undefined,
+            schema ? [tableName, schema] : [tableName],
+        );
+        return this.parseWarehouseCatalog(rows, mapFieldType);
     }
 }